Enhance the child node creation process
Logging can be turned on when a node is created. The `logging_on` parameter when set to any value turns on the pico labs logger when the child is created.

This setups the system_manager to better install rulesets and prepares the way to introduce picos via subscriptions to start the network.